+++ modules/apps/attack.js
@@ -1,5 +1,13 @@
 import { HMPrompt } from './prompt.js';
-import { HMCONST } from '../sys/constants.js';
+import { HMCONST, HMTABLES } from '../sys/constants.js';
+
+function getSpeed(ranged, wData, specialMove=0) {
+    const {spd, jspd} = wData.bonus.total;
+    if (!ranged) return {melee: Number(specialMove) === HMCONST.SPECIAL.JAB ? jspd : spd};
+
+    const {timing} = wData.ranged;
+    return HMTABLES.weapons.ranged.timing.declare(timing, spd);
+}
 
 export class AttackPrompt extends HMPrompt {
     static get defaultOptions() {
@@ -14,12 +22,15 @@ export class AttackPrompt extends HMPrompt {
         const weapon = dialogData.weapons[0];
         const capList = this.getCapList(weapon.capabilities, dialogData?.caller);
 
+        const wData = weapon.data.data;
+        const ranged = wData.ranged.checked;
+        const spd = getSpeed(ranged, wData);
+
         mergeObject(this.dialogData, {
-            caller: dialogData?.caller,
             capList,
             specialMove: 0,
-            ranged: weapon.data.data.ranged.checked,
-            spd: weapon.data.data.bonus.total.spd,
+            ranged,
+            spd,
             widx: 0,
             range: 0,
             advance: dialogData.inCombat,
@@ -27,27 +38,26 @@ export class AttackPrompt extends HMPrompt {
     }
 
     update(options) {
-        const {weapons, widx, caller} = this.dialogData;
-        const weapon   = weapons[widx];
-        const wData    = weapon.data.data;
-        const {ranged} = wData;
-        const spd = Number(this.dialogData.specialMove) === HMCONST.SPECIAL.JAB
-                        ? wData.bonus.total.jspd
-                        : wData.bonus.total.spd;
+        const {weapons, widx, caller, specialMove} = this.dialogData;
+        const weapon  = weapons[widx];
+        const wData   = weapon.data.data;
+        const ranged  = wData.ranged.checked;
 
-        this.dialogData.ranged = ranged.checked;
-        this.dialogData.spd = spd;
+        this.dialogData.ranged = ranged;
+        this.dialogData.spd = getSpeed(ranged, wData, specialMove);
         this.dialogData.capList = this.getCapList(weapons[widx].capabilities, caller);
         super.update(options);
     }
 
     get dialogResp() {
+        const {button, spd} = this.dialogData;
         const dialogResp = {
             widx: this.dialogData.widx,
             specialMove: Number(this.dialogData.specialMove),
             range: Number(this.dialogData.range),
             bonus: parseInt(this.dialogData.bonus, 10) || 0,
-            advance: this.dialogData.advance ? this.dialogData.spd : false,
+            advance: this.dialogData.advance ? spd[button] : false,
+            button,
         };
         return dialogResp;
     }