+++ modules/item/weapon-item.js
@@ -26,7 +26,7 @@ export class HMWeaponItem extends HMItem {
                                                        && a.invstate === 'equipped');
 
         if (itemData.innate) itemData.state = HMCONST.ITEM_STATE.INNATE;
-        // HACK: This belongs in item-sheet.js, which needs a refactor.
+        // TODO: Some of this can be relocated to weapon-item-sheet.js.
         const {reach} = this.data.data;
         const offset = this.parent.data.data.bonus.total.reach || 0;
         itemData.adjReach = Math.max(reach + offset, 0) || 0;
@@ -172,8 +172,11 @@ export class HMWeaponItem extends HMItem {
         const ranged = dialogResp.context.data.data.ranged.checked;
         if (ranged) dataset.dialog = 'ratk';
 
-        const rollMgr = new HMRollMgr();
-        const roll = await rollMgr.getRoll(dataset, dialogResp);
+        let roll;
+        if (dialogResp.resp.button !== 'declare') {
+            const rollMgr = new HMRollMgr();
+            roll = await rollMgr.getRoll(dataset, dialogResp);
+        }
 
         const chatMgr = new HMChatMgr();
         const card = await chatMgr.getCard({roll, dataset, dialogResp});