+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### Added
 - Dialog buttons which advance clock will display seconds to be advanced.
+- Ranged weapons include a full timing breakdown.
 ### Fixed
 - Cleared token dependency on actor embedded items.